dcp-815 separate supplementary_file_link.py
