+++ exporter/graph/crawler.py
@@ -4,12 +4,13 @@ from functools import reduce
 from operator import iconcat
 from typing import List, Iterable, Optional, Callable
 
-from exporter.graph.experiment_graph import ExperimentGraph, SupplementaryFileLink
+from exporter.graph.experiment_graph import ExperimentGraph
 from exporter.graph.input import Input
 from exporter.graph.output import Output
 from exporter.graph.process_link import ProcessLink
 from exporter.graph.protocol_link import ProtocolLink
 from exporter.graph.supplementary_file import SupplementaryFile
+from exporter.graph.supplementary_file_link import SupplementaryFileLink
 from exporter.graph.supplemented_entity import SupplementedEntity
 from exporter.metadata.resource import MetadataResource
 from exporter.metadata.service import MetadataService
