+++ tests/exporter/graph/test_experiment_graph.py
@@ -1,7 +1,8 @@
 from unittest import TestCase
 
-from exporter.graph.experiment_graph import LinkSet, SupplementaryFileLink
+from exporter.graph.experiment_graph import LinkSet
 from exporter.graph.process_link import ProcessLink
+from exporter.graph.supplementary_file_link import SupplementaryFileLink
 from exporter.graph.supplemented_entity import SupplementedEntity
